requirements.txt
python-dotenv==0.12.0
alembic==1.4.1
aniso8601==8.0.0
attrs==19.3.0
click==7.1.1
Flask==1.1.1
Flask-JWT-Extended==3.24.1
flask-restplus==0.13.0
Flask-SQLAlchemy-Session==1.1
gunicorn==20.0.4
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
Mako==1.1.2
MarkupSafe==1.1.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
psycopg2==2.8.4
py==1.8.1
PyJWT==1.7.1
pyparsing==2.4.6
pyrsistent==0.15.7
pytest==5.4.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.15
wcwidth==0.1.8
Werkzeug==0.16.1
zipp==3.1.0
